refactor: NoteWarpListを分離
