+++ apps/client/src/pages/note-list/ui/page.tsx
@@ -1,36 +1,15 @@
-import { cx, style } from "@acalyle/css";
-import { Intersection, TextInput } from "@acalyle/ui";
-import { useCallback, useDeferredValue, useState } from "react";
-import { type BookRef, useBookByRef } from "~/entities/book";
-import { $noteConnection, type NoteId, NoteOverview } from "~/entities/note";
+import { TextInput } from "@acalyle/ui";
+import { useDeferredValue, useState } from "react";
+import type { BookRef } from "~/entities/book";
 import { NoteCreateButton } from "~/features/create-note";
 import { NoteModalContainer } from "~/features/note-modal";
-import { MIN_NOTE_WIDTH } from "~/note/ui/constants";
-import { useConnection } from "~/shared/graphql";
+import { NoteWarpList } from "./warp-list";
 
 export const NoteListPage: React.FC<{
   bookRef: BookRef;
 }> = ({ bookRef }) => {
   const [query, setQuery] = useState("");
   const deferredQuery = useDeferredValue(query);
-
-  // FIXME: non-null ではない
-  const bookId = useBookByRef(bookRef)!.id;
-  const { nodeIds } = useConnection(
-    $noteConnection(bookId, `-@relate:* ${deferredQuery}`),
-  );
-
-  const onIntersection = useCallback(
-    (entry: IntersectionObserverEntry) => {
-      if (entry.isIntersecting) {
-        void $noteConnection(
-          bookId,
-          `-@relate:* ${deferredQuery}`,
-        ).loadNextPage();
-      }
-    },
-    [bookId, deferredQuery],
-  );
 
   return (
     <div>
@@ -42,23 +21,7 @@ export const NoteListPage: React.FC<{
         />
         <NoteCreateButton bookRef={bookRef} />
       </div>
-      <div
-        className={cx(
-          ":uno: grid gap-y-4 gap-x-5 grid-auto-rows-48",
-          style({
-            gridTemplateColumns: `repeat(auto-fill, minmax(${MIN_NOTE_WIDTH}, 1fr))`,
-          }),
-        )}
-      >
-        {nodeIds.map(noteId => (
-          <NoteOverview
-            key={noteId}
-            bookId={bookId}
-            noteId={noteId as string as NoteId}
-          />
-        ))}
-      </div>
-      <Intersection onIntersection={onIntersection} rootMargin="25% 0px" />
+      <NoteWarpList bookRef={bookRef} query={`-@relate:* ${deferredQuery}`} />
       <NoteModalContainer />
     </div>
   );
